pyproject.toml
[project]
name = "granked-data-pipeline"
version = "0.1.0"
description = "Data pipeline for Granked"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "langdetect>=1.0.9",
    "llama-cpp-python>=0.3.20",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "requests>=2.33.1",
]
